Rename groupe_parser to group_parser and info_of_test_parser to test_info_parser
