+++ openscap_report/scap_results_parser/parsers/__init__.py
@@ -2,9 +2,9 @@
 # SPDX-License-Identifier: LGPL-2.1-or-later
 
 from .full_text_parser import FullTextParser
-from .groupe_parser import GroupParser
-from .info_of_test_parser import InfoOfTest
+from .group_parser import GroupParser
 from .oval_definition_parser import OVALDefinitionParser
 from .remediation_parser import RemediationParser
 from .report_parser import ReportParser
 from .rule_parser import RuleParser
+from .test_info_parser import InfoOfTest
