+++ openscap_report/scap_results_parser/parsers/group_parser.py
